requirements.txt
coverage
nose
xlrd
pyyaml
